+++ app/src/main/java/com/alphawallet/app/interact/CreateTransactionInteract.java
@@ -28,25 +28,25 @@ public class CreateTransactionInteract
         this.transactionRepository = transactionRepository;
     }
 
-    public Single<SignaturePair> sign(Wallet wallet, MessagePair messagePair, int chainId)
+    public Single<SignaturePair> sign(Wallet wallet, MessagePair messagePair, long chainId)
     {
         return transactionRepository.getSignature(wallet, messagePair, chainId)
                 .map(sig -> new SignaturePair(messagePair.selection, sig.signature, messagePair.message));
     }
 
-    public Single<SignatureFromKey> sign(Wallet wallet, Signable message, int chainId)
+    public Single<SignatureFromKey> sign(Wallet wallet, Signable message, long chainId)
     {
         return transactionRepository.getSignature(wallet, message, chainId);
     }
 
-    public Single<String> create(Wallet from, String to, BigInteger subunitAmount, BigInteger gasPrice, BigInteger gasLimit, byte[] data, int chainId)
+    public Single<String> create(Wallet from, String to, BigInteger subunitAmount, BigInteger gasPrice, BigInteger gasLimit, byte[] data, long chainId)
     {
         return transactionRepository.createTransaction(from, to, subunitAmount, gasPrice, gasLimit, data, chainId)
                                              .subscribeOn(Schedulers.computation())
                                              .observeOn(AndroidSchedulers.mainThread());
     }
 
-    public Single<TransactionData> createWithSig(Wallet from, Web3Transaction web3Tx, int chainId)
+    public Single<TransactionData> createWithSig(Wallet from, Web3Transaction web3Tx, long chainId)
     {
         return transactionRepository.createTransactionWithSig(from, web3Tx.recipient.toString(), web3Tx.value,
                         web3Tx.gasPrice, web3Tx.gasLimit, web3Tx.nonce,
@@ -55,21 +55,21 @@ public class CreateTransactionInteract
                                          .observeOn(AndroidSchedulers.mainThread());
     }
 
-    public Single<TransactionData> createWithSig(Wallet from, String to, BigInteger subunitAmount, BigInteger gasPrice, BigInteger gasLimit, byte[] data, int chainId)
+    public Single<TransactionData> createWithSig(Wallet from, String to, BigInteger subunitAmount, BigInteger gasPrice, BigInteger gasLimit, byte[] data, long chainId)
     {
         return transactionRepository.createTransactionWithSig(from, to, subunitAmount, gasPrice, gasLimit, -1, data, chainId)
                 .subscribeOn(Schedulers.computation())
                 .observeOn(AndroidSchedulers.mainThread());
     }
 
-    public Single<TransactionData> createWithSig(Wallet from, BigInteger gasPrice, BigInteger gasLimit, String data, int chainId)
+    public Single<TransactionData> createWithSig(Wallet from, BigInteger gasPrice, BigInteger gasLimit, String data, long chainId)
     {
         return transactionRepository.createTransactionWithSig(from, gasPrice, gasLimit, data, chainId)
                                          .subscribeOn(Schedulers.computation())
                                          .observeOn(AndroidSchedulers.mainThread());
     }
 
-    public Single<String> resend(Wallet from, BigInteger nonce, String to, BigInteger subunitAmount, BigInteger gasPrice, BigInteger gasLimit, byte[] data, int chainId)
+    public Single<String> resend(Wallet from, BigInteger nonce, String to, BigInteger subunitAmount, BigInteger gasPrice, BigInteger gasLimit, byte[] data, long chainId)
     {
         return transactionRepository.resendTransaction(from, to, subunitAmount, nonce, gasPrice, gasLimit, data, chainId)
                 .subscribeOn(Schedulers.computation())
@@ -81,7 +81,7 @@ public class CreateTransactionInteract
         transactionRepository.removeOverridenTransaction(wallet, oldTxHash);
     }
 
-    public Single<TransactionData> signTransaction(Wallet wallet, Web3Transaction w3tx, int chainId)
+    public Single<TransactionData> signTransaction(Wallet wallet, Web3Transaction w3tx, long chainId)
     {
         return transactionRepository.getSignatureForTransaction(wallet, w3tx, chainId);
     }
