+++ app/src/main/java/com/alphawallet/app/ui/ImportTokenActivity.java
@@ -80,7 +80,7 @@ public class ImportTokenActivity extends BaseActivity implements View.OnClickLis
     private TextView importTxt;
 
     private LinearLayout costLayout;
-    private int chainId = 0;
+    private long chainId = 0;
     private boolean usingFeeMaster = false;
     private String paymasterUrlPrefix = "https://paymaster.stormbird.sg/api/";
     private final String TAG = "ITA";
