+++ app/src/main/java/com/alphawallet/app/viewmodel/CustomNetworkViewModel.java
@@ -18,11 +18,11 @@ public class CustomNetworkViewModel extends BaseViewModel
         this.ethereumNetworkRepository = ethereumNetworkRepository;
     }
 
-    public void addNetwork(String name, String rpcUrl, int chainId, String symbol, String blockExplorerUrl, String explorerApiUrl, boolean isTestnet, Integer oldChainId) {
+    public void addNetwork(String name, String rpcUrl, long chainId, String symbol, String blockExplorerUrl, String explorerApiUrl, boolean isTestnet, Long oldChainId) {
         this.ethereumNetworkRepository.addCustomRPCNetwork(name, rpcUrl, chainId, symbol, blockExplorerUrl, explorerApiUrl, isTestnet, oldChainId);
     }
 
-    public EthereumNetworkRepositoryType.NetworkInfoExt getNetworkInfo(int chainId) {
+    public EthereumNetworkRepositoryType.NetworkInfoExt getNetworkInfo(long chainId) {
         return this.ethereumNetworkRepository.getNetworkInfoExt(chainId);
     }
 }
