+++ lib/src/main/java/com/alphawallet/token/entity/AttributeInterface.java
@@ -16,7 +16,7 @@ public interface AttributeInterface
 
     String getWalletAddr();
 
-    default long getLastTokenUpdate(int chainId, String address) { return 0; };
+    default long getLastTokenUpdate(long chainId, String address) { return 0; };
     default Attribute fetchAttribute(ContractInfo origin, String attributeName) { return null; };
     default TokenScriptResult.Attribute fetchAttrResult(ContractAddress origin, Attribute attr, BigInteger tokenId) { return null; };
 }
