+++ lib/src/main/java/com/alphawallet/token/entity/ContractInfo.java
@@ -11,9 +11,9 @@ import java.util.Map;
 public class ContractInfo
 {
     public final String contractInterface;
-    public final Map<Integer, List<String>> addresses = new HashMap<>();
+    public final Map<Long, List<String>> addresses = new HashMap<>();
 
-    public ContractInfo(String contractType, Map<Integer, List<String>> addresses)
+    public ContractInfo(String contractType, Map<Long, List<String>> addresses)
     {
         this.contractInterface = contractType;
         this.addresses.putAll(addresses);
@@ -24,13 +24,13 @@ public class ContractInfo
         this.contractInterface = contractType;
     }
 
-    public boolean hasContractTokenScript(int chainId, String address)
+    public boolean hasContractTokenScript(long chainId, String address)
     {
         List<String> addrs = addresses.get(chainId);
         return addrs != null && addrs.contains(address);
     }
 
-    public int getfirstChainId()
+    public long getfirstChainId()
     {
         if (addresses.keySet().size() > 0) return addresses.keySet().iterator().next();
         else return 0;
@@ -38,7 +38,7 @@ public class ContractInfo
 
     public String getFirstAddress()
     {
-        int chainId = getfirstChainId();
+        long chainId = getfirstChainId();
         return addresses.get(chainId).size() > 0 ? addresses.get(chainId).get(0) : "";
     }
 }
